a bug

+++ generate_data.py
@@ -38,7 +38,7 @@ def combine_spectrum(paths, max_ndvi=None):
     raw_spectrum = {}
     for band, path in paths.items():
         # prone to bug/ check whether cv2 read int16 correctly.
-        raw_spectrum[band] = cv2.resize(cv2.imread(path, cv2.IMREAD_ANYDEPTH).astype(np.float16), dsize=SIZE)
+        raw_spectrum[band] = cv2.resize(cv2.imread(path, cv2.IMREAD_ANYDEPTH).astype(np.uint16), dsize=SIZE)
     # ignore scl/tci/wvp just to keep it raw
     
     #ndvi = features.ndvi(raw_spectrum)
